+++ alfresco-health-processor-platform/src/main/java/eu/xenit/alfresco/healthprocessor/indexing/txnaggregation/TransactionAggregationIndexingStrategyState.java
@@ -1,4 +1,4 @@
-package eu.xenit.alfresco.healthprocessor.indexing.threshold;
+package eu.xenit.alfresco.healthprocessor.indexing.txnaggregation;
 
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.binder.MeterBinder;
@@ -13,7 +13,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ThresholdIndexingStrategyState implements MeterBinder {
+public class TransactionAggregationIndexingStrategyState implements MeterBinder {
 
     private long currentTransactionId = -1;
     private long maxTransactionId = -1;
@@ -35,9 +35,9 @@ public class ThresholdIndexingStrategyState implements MeterBinder {
 
     @Override
     public void bindTo(MeterRegistry registry) {
-        registry.gauge("eu.xenit.alfresco.healthprocessor.indexing.threshold.current-transaction-id", this, ThresholdIndexingStrategyState::getCurrentTransactionId);
-        registry.gauge("eu.xenit.alfresco.healthprocessor.indexing.threshold.max-transaction-id", this, ThresholdIndexingStrategyState::getMaxTransactionId);
-        registry.gauge("eu.xenit.alfresco.healthprocessor.indexing.threshold.running-transaction-mergers", this, ThresholdIndexingStrategyState::getRunningTransactionMergers);
+        registry.gauge("eu.xenit.alfresco.healthprocessor.indexing.threshold.current-transaction-id", this, TransactionAggregationIndexingStrategyState::getCurrentTransactionId);
+        registry.gauge("eu.xenit.alfresco.healthprocessor.indexing.threshold.max-transaction-id", this, TransactionAggregationIndexingStrategyState::getMaxTransactionId);
+        registry.gauge("eu.xenit.alfresco.healthprocessor.indexing.threshold.running-transaction-mergers", this, TransactionAggregationIndexingStrategyState::getRunningTransactionMergers);
         registry.gauge("eu.xenit.alfresco.healthprocessor.indexing.threshold.queued-transaction-batches", transactionBatchesQueueSize, AtomicInteger::get);
     }
 }